+++ src/lib/libcmd/rm.c
@@ -72,7 +72,7 @@ static const char usage[] =
 #define isempty(f)	(!((f)->fts_number&RM_ENTRY))
 #define nonempty(f)	((f)->fts_parent->fts_number|=RM_ENTRY)
 #define pathchunk(n)	roundof(n,1024)
-#define retry(f)	((f)->fts_number=((f)->fts_statp->st_nlink<<1))
+#define retry(f)	((f)->fts_number=((long)(((f)->fts_statp->st_nlink<<1))))
 
 typedef struct State_s			/* program state		*/
 {
